#0: Mochi: Rename and move modules
