+++ models/experimental/mochi/tests/dit/test_mlp.py
@@ -3,15 +3,13 @@ import pytest
 from loguru import logger
 import os
 import ttnn
-from models.experimental.mochi.ff import TtFeedForward
-from models.demos.llama3.tt.model_config import TtModelArgs
+from models.experimental.mochi.tt.dit.mlp import FeedForward as TtFeedForward
 from models.utility_functions import (
-    comp_pcc,
     comp_allclose,
 )
 from models.utility_functions import skip_for_grayskull
 from genmo.mochi_preview.dit.joint_model.layers import FeedForward as RefFeedForward
-from models.experimental.mochi.common import get_mochi_dir, get_cache_path, compute_metrics
+from models.experimental.mochi.tt.common import get_mochi_dir, get_cache_path, compute_metrics
 
 
 @torch.no_grad()
